+++ src/PrestaShopBundle/Resources/views/Admin/Improve/International/Translations/overview.html.twig
@@ -25,34 +25,34 @@
 {% extends '@PrestaShop/Admin/layout.html.twig' %}
 
 {% block content %}
-    <div id="translations-app"></div>
+  <div id="translations-app"></div>
 {% endblock %}
 
 {% block javascripts %}
-    {{ parent() }}
+  {{ parent() }}
 
-    <script>
-        var data = {
-            baseUrl: '{{ app.request.getBasePath() }}',
-            locale: '{{ app.request.query.get('locale') }}',
-            type: '{{ app.request.query.get('type') }}',
-            currentTheme: {{ theme.name|json_encode|raw }},
-            selected: '{{ app.request.query.get('selected') }}',
-            domainsTreeUrl: '{{ url('api_translation_domains_tree', {
-              'lang' : app.request.query.get('lang'),
-              'type' : app.request.query.get('type'),
-              'selected': app.request.query.get('selected')
-            }) }}',
-            translationUrl: '{{ url('api_i18n_translations_list', {'page' : 'international'}) }}',
-            internationalUrl: '{{ getAdminLink("AdminLocalization")|raw }}',
-            translationsUrl: '{{ getAdminLink("AdminTranslations")|raw }}'
-        }
-    </script>
+  <script>
+    var data = {
+      baseUrl: '{{ app.request.getBasePath() }}',
+      locale: '{{ app.request.query.get('locale') }}',
+      type: '{{ app.request.query.get('type') }}',
+      currentTheme: {{ theme.name|json_encode|raw }},
+      selected: '{{ app.request.query.get('selected') }}',
+      domainsTreeUrl: '{{ url('api_translation_domains_tree', {
+        'lang' : app.request.query.get('lang'),
+        'type' : app.request.query.get('type'),
+        'selected': app.request.query.get('selected')
+      }) }}',
+      translationUrl: '{{ url('api_i18n_translations_list', {'page' : 'international'}) }}',
+      internationalUrl: '{{ getAdminLink("AdminLocalization")|raw }}',
+      translationsUrl: '{{ getAdminLink("AdminTranslations")|raw }}'
+    }
+  </script>
 
-    {% if webpack_server %}
-        <script src="http://localhost:8080/translations.bundle.js"></script>
-    {% else %}
-        <script src="{{ asset('themes/new-theme/public/translations.bundle.js') }}"></script>
-    {% endif %}
+  {% if webpack_server %}
+    <script src="http://localhost:8080/translations.bundle.js"></script>
+  {% else %}
+    <script src="{{ asset('themes/new-theme/public/translations.bundle.js') }}"></script>
+  {% endif %}
 
 {% endblock %}
